requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-filter
pillow
drf-yasg
djangorestframework-simplejwt
dj-config-url
django-cors-headers
redis
celery
django-allauth